# vw_gateway/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /video/list:
        get:
            tags:
                - VideoInfo
            operationId: VideoInfo_GetVideoList
            parameters:
                - name: class
                  in: query
                  schema:
                    type: array
                    items:
                        type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.api.v1.video.videoinfo.GetVideoListResp'
...
    /video/{videoId}/info:
        get:
            tags:
                - VideoInfo
            operationId: VideoInfo_GetVideoInfo
            parameters:
                - name: videoId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.api.v1.video.videoinfo.GetVideoInfoResp'
...
components:
    schemas:
...
        gateway.api.v1.video.videoinfo.GetVideoInfoResp:
            type: object
            properties:
                title:
                    type: string
        gateway.api.v1.video.videoinfo.GetVideoListResp:
            type: object
            properties:
                req:
                    type: array
                    items:
                        $ref: '#/components/schemas/gateway.api.v1.video.videoinfo.VideoMetaInfo'
...
        gateway.api.v1.video.videoinfo.VideoMetaInfo:
            type: object
            properties:
                videoId:
                    type: string
                authorId:
                    type: string
                authorName:
                    type: string
                title:
                    type: string
                path:
                    type: string
                className:
                    type: string
                tags:
                    type: string
                hot:
                    type: string
                createdAt:
                    type: string
                    format: date-time
                records:
                    $ref: '#/components/schemas/gateway.api.v1.video.videoinfo.VideoMetaInfo_Records'
                durations:
                    type: string
                coverPath:
                    type: string
        gateway.api.v1.video.videoinfo.VideoMetaInfo_Records:
            type: object
            properties:
                cntBarrages:
                    type: integer
                    format: uint32
                cntShares:
                    type: integer
                    format: uint32
                cntViewed:
                    type: integer
                    format: uint32
                cntFavorites:
                    type: integer
                    format: uint32
...